+++ src/listaA/Main.java
@@ -7,7 +7,7 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner scanner = new Scanner(System.in);
 		System.out.println("Qual exercício você deseja fazer?");
-		int opt = scanner.nextInt();
+		byte opt = scanner.nextByte();
 		Exercicios exercise = new Exercicios(opt);
 		exercise.execute();
 		scanner.close();
